refactor(download): refactoring TCX-stuff
