+++ src/main/kotlin/org/devshred/gpstools/web/GpxMergeController.kt
@@ -6,8 +6,7 @@ import org.devshred.gpstools.domain.IOService
 import org.devshred.gpstools.domain.NotFoundException
 import org.devshred.gpstools.domain.StoredFile
 import org.devshred.gpstools.domain.gpx.buildGpx
-import org.devshred.gpstools.domain.proto.gpxToProtobufInputStream
-import org.devshred.gpstools.domain.proto.protoInputStreamResourceToProtoGpsContainer
+import org.devshred.gpstools.domain.proto.ProtoService
 import org.devshred.gpstools.domain.proto.toGpx
 import org.slf4j.LoggerFactory
 import org.springframework.http.ResponseEntity
@@ -23,7 +22,11 @@ import java.util.UUID
 @CrossOrigin(origins = ["*"], maxAge = 3600)
 @RestController
 @RequestMapping("/merge")
-class GpxMergeController(private val store: FileStore, private val ioService: IOService) {
+class GpxMergeController(
+    private val store: FileStore,
+    private val ioService: IOService,
+    private val protoService: ProtoService,
+) {
     private val log = LoggerFactory.getLogger(javaClass)
 
     @ResponseBody
@@ -46,16 +49,16 @@ class GpxMergeController(private val store: FileStore, private val ioService: IO
         var trackName: String? = null
         fileIds.forEachIndexed { index, uuid ->
             log.info("About to merge $uuid.")
-            val gpsContainer =
-                protoInputStreamResourceToProtoGpsContainer(ioService.getAsStream(store.get(uuid).storageLocation))
-            allWayPoints.addAll(gpsContainer.wayPointsList.map { toGpx(it) })
-            allTrackPoints.addAll(gpsContainer.track.wayPointsList.map { toGpx(it) })
-            if (index == 0 && gpsContainer.name.isNotEmpty()) {
-                trackName = gpsContainer.name
+            val protoGpsContainer =
+                protoService.readProtoGpsContainer(store.get(uuid).storageLocation)
+            allWayPoints.addAll(protoGpsContainer.wayPointsList.map { it.toGpx() })
+            allTrackPoints.addAll(protoGpsContainer.track.wayPointsList.map { it.toGpx() })
+            if (index == 0 && protoGpsContainer.name.isNotEmpty()) {
+                trackName = protoGpsContainer.name
             }
         }
         val gpx = buildGpx(trackName, allWayPoints, allTrackPoints)
-        val protoStream = gpxToProtobufInputStream(gpx)
+        val protoStream = protoService.gpxToProtobufInputStream(gpx)
         val protoFile = ioService.createTempFile(protoStream, "merged.gpx")
         store.put(protoFile.id, protoFile)
