+++ src/test/resources/data/full.tcx
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<TrainingCenterDatabase xmlns="http://www.garmin.com/xmlschemas/TrainingCenterDatabase/v2">
+    <Courses>
+        <Course>
+            <Name>Happy Path Example</Name>
+            <Lap>
+                <TotalTimeSeconds>28771.0</TotalTimeSeconds>
+                <DistanceMeters>64757.0470678665</DistanceMeters>
+                <BeginPosition>
+                    <LatitudeDegrees>36.72100500</LatitudeDegrees>
+                    <LongitudeDegrees>-4.41088200</LongitudeDegrees>
+                </BeginPosition>
+                <EndPosition>
+                    <LatitudeDegrees>36.73361890</LatitudeDegrees>
+                    <LongitudeDegrees>-3.68807099</LongitudeDegrees>
+                </EndPosition>
+                <Intensity>Active</Intensity>
+            </Lap>
+            <Track>
+                <Trackpoint>
+                    <Time>2010-01-01T00:00:00Z</Time>
+                    <Position>
+                        <LatitudeDegrees>36.72100500</LatitudeDegrees>
+                        <LongitudeDegrees>-4.41088200</LongitudeDegrees>
+                    </Position>
+                    <AltitudeMeters>14.000000</AltitudeMeters>
+                    <DistanceMeters>0.000</DistanceMeters>
+                </Trackpoint>
+                <Trackpoint>
+                    <Time>2010-01-01T03:16:04Z</Time>
+                    <Position>
+                        <LatitudeDegrees>36.74881700</LatitudeDegrees>
+                        <LongitudeDegrees>-4.07262399</LongitudeDegrees>
+                    </Position>
+                    <AltitudeMeters>3.000000</AltitudeMeters>
+                    <DistanceMeters>30370.327</DistanceMeters>
+                </Trackpoint>
+                <Trackpoint>
+                    <Time>2010-01-01T07:59:31Z</Time>
+                    <Position>
+                        <LatitudeDegrees>36.73361890</LatitudeDegrees>
+                        <LongitudeDegrees>-3.68807099</LongitudeDegrees>
+                    </Position>
+                    <AltitudeMeters>11.000000</AltitudeMeters>
+                    <DistanceMeters>64757.047</DistanceMeters>
+                </Trackpoint>
+            </Track>
+            <CoursePoint>
+                <Name>Highlight</Name>
+                <Time>2010-01-01T03:16:04Z</Time>
+                <Position>
+                    <LatitudeDegrees>36.74881700</LatitudeDegrees>
+                    <LongitudeDegrees>-4.07262399</LongitudeDegrees>
+                </Position>
+                <PointType>Food</PointType>
+            </CoursePoint>
+        </Course>
+    </Courses>
+</TrainingCenterDatabase>
